requirements.txt
torch>=2.1.0
numpy>=1.26.0
Pillow>=10.0.0
kornia>=0.7.0
tensorboard>=2.14.0
tqdm>=4.66.0
gradio>=4.0.0
